list users:

+++ db/schema.rb
@@ -26,7 +26,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_09_02_204550) do
     t.string "filename", null: false
     t.string "content_type"
     t.text "metadata"
-    t.bigint "byte_size", null: false
+    t.integer "byte_size", null: false
     t.string "checksum"
     t.datetime "created_at", precision: nil, null: false
     t.string "service_name", null: false
